lol3

+++ test_companion.py
@@ -1,5 +1,5 @@
 def test_companion(test_str: str) -> dict[str, str]:
     pass
 
-def grade_answer(question: str, answer: str, correct_answer: str) -> tuple[float, str]:
+def grade_answer(question: str, answer: str, correct_answer: str) -> tuple[int, str]:
     pass
